+++ src/plugin-info.ts
@@ -1,5 +1,5 @@
 /* File will be overwritten by bin/release.sh! */
 export const PLUGIN_INFO = {
-  "pluginVersion": "1.6.5",
-  "pluginReleasedAt": "2024-10-09T13:17:57+0200"
+  "pluginVersion": "1.7.0",
+  "pluginReleasedAt": "2025-01-30T12:09:48+0100"
 }