+++ src/routes/dataview.ts
@@ -98,6 +98,23 @@ async function executeDataviewQuery(
     return success({ data: dqlValuesMapper(dataview, res.value.values[0]) });
   }
 
+  // For LIST queries, DV will return a two-dimensional array instead of a one-
+  // dimensional one *if* one of the queried files returns more than one hit.
+  // This is inconsistent, and AFO will nope out. So we'll make it consistent.
+  //
+  // Example: If you query for an inline field (`whatever::`), and one file
+  // contains two of this field, e.g. `whatever:: something 1` and
+  // `whatever:: something 2`, while another file contains just one
+  // (e.g., `whatever:: something 3`), DV will return:
+  //
+  //     [
+  //       ["something 1", "something 2"],
+  //       "something 3"
+  //     ]
+  if (type === "list") {
+    res.value.values = res.value.values.map((v: any) => Array.isArray(v) ? v : [v])
+  }
+
   return success({ data: dqlValuesMapper(dataview, res.value.values) });
 }